Iniciante

+++ URI/1021.cpp
@@ -7,7 +7,7 @@ int main() {
 	double vr; // Valor real
 	scanf("%lf", &vr);
 	int notas[6] = {100, 50, 20, 10, 5, 2}; // Valor das cédulas
-	double moedas[6] = {100, 50, 25, 10, 5, 1}; // Valor dos centavos (em int para usar o operador %)
+	int moedas[6] = {100, 50, 25, 10, 5, 1}; // Valor dos centavos (em int para usar o operador %)
 	int vr_notas = (int) vr; // Pega apenas o valor em cédulas do valor real
 	int vr_moedas = (int) (100.0 * (vr - (double) vr_notas)); // Pega o valor inteiro de centavos
 	printf("NOTAS:\n");
